# added benchmark for RecordGenerator.insertRR