- fixes #408 

Micro-benchmark (take with a grain of salt) shows about 20x gain in performance.

Before improvements in #395:
```
BenchmarkInsertRR         500000              2394 ns/op
BenchmarkInsertRR         500000              2383 ns/op
BenchmarkInsertRR         500000              2402 ns/op
BenchmarkInsertRR         500000              2439 ns/op
BenchmarkInsertRR         500000              2423 ns/op
BenchmarkInsertRR         500000              2412 ns/op
```

After:
```
BenchmarkInsertRR       10000000               139 ns/op
BenchmarkInsertRR       10000000               147 ns/op
BenchmarkInsertRR       10000000               140 ns/op
BenchmarkInsertRR       10000000               139 ns/op
BenchmarkInsertRR       10000000               138 ns/op
BenchmarkInsertRR       10000000               147 ns/op
```