@sargun 2016-03-07
Couple notes. Addressable by just throwing comments into the code to explain this code to my brain in a month. But, great job. It's wonderful to see benchmark, and analysis.